Re-introduce some platform-specific casts and allow the clippy::unnecessary_casts lint

+++ lib/compiler-singlepass/src/machine_arm64.rs
@@ -234,7 +234,8 @@ impl MachineARM64 {
                         temps.push(tmp);
                         tmp
                     };
-                    self.assembler.emit_mov_imm(Location::GPR(tmp), val)?;
+                    self.assembler
+                        .emit_mov_imm(Location::GPR(tmp), val as u64)?;
                     Ok(Location::GPR(tmp))
                 }
             }
@@ -388,7 +389,8 @@ impl MachineARM64 {
                         CompileError::Codegen("singlepass cannot acquire temp simd".to_owned())
                     })?;
                     temps.push(tmp);
-                    self.assembler.emit_mov_imm(Location::GPR(gpr), val)?;
+                    self.assembler
+                        .emit_mov_imm(Location::GPR(gpr), val as u64)?;
                     self.assembler
                         .emit_mov(sz, Location::GPR(gpr), Location::SIMD(tmp))?;
                     self.release_gpr(gpr);
